Trapezoid and Circle

+++ src/main/java/Rectangle.java
@@ -6,9 +6,9 @@ public class Rectangle implements Figure {
     public double calculateArea() {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Podaj długość boku a:");
-        int a = scanner.nextInt();
+        double a = scanner.nextInt();
         System.out.println("Podaj długość boku b:");
-        int b = scanner.nextInt();;
+        double b = scanner.nextInt();;
         return a*b;
     }
 }
